\name{H27MPJ}
\docType{data}
\alias{H27MPJ}
\title{Tentative model population of Japan in Heisei 27 (2015)}
\description{
 The data gives the age-class specific model population of Japan
 in smoothed Heisei 27 (2015) to calculate directly adjusted mortality rate.
}
\usage{ H27MPJ }
\format{ A named vector containing 21 observations, where names show age-classes. }
\source{ \url{https://www.mhlw.go.jp/content/12601000/000638712.pdf} }
\references{
 Tamura K. (2008) How do we die?: death date from vital statistics of the
 Japanese population. \emph{The Waseda study of politics and public law},
 87: 27-57.
}
\keyword{datasets}
