\name{S60MPJ}
\docType{data}
\alias{S60MPJ}
\title{Model population of Japan in Showa 60 (1985)}
\description{
 The data gives the age-class (by five) specific model population of Japan
 in Showa 60 (1985) to calculate directly adjusted mortality rate.
}
\usage{ S60MPJ }
\format{ A vector containing 18 observations. }
\source{ \url{https://www.mhlw.go.jp/toukei/list/dl/81-1b1.pdf}, page 55. }
\references{
 Tamura K. (2008) How do we die?: death date from vital statistics of the
 Japanese population. \emph{The Waseda study of politics and public law},
 87: 27-57.
}
\keyword{datasets}
